Py-Docker/requirements.txt
psycopg2-binary
boto3
flask
